+++ CHANGELOG.md
@@ -4,6 +4,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [5.3.0] - 2024-03-11
+
+### Changed
+- Updated dependencies
+- Updated min dart sdk to `3.3.0`
+
+### Fixed
+- `analyze` task now honors `error-level` setting and will only fail if the analyzer fails as well (#30)
+  - It will still log info/warning messages
+
+### Removed
+- Removed the `scan-mode` option of the `analyze` task (#30)
+  - It will now always scan the whole repository for issues
+
 ## [5.2.1] - 2023-08-16
 ### Changed
 - Update dependencies
@@ -196,6 +210,7 @@ higher version from the lockfile
 - Initial release
 - Automatic deployment
 
+[5.3.0]: https://github.com/Skycoder42/dart_pre_commit/compare/v5.2.1...v5.3.0
 [5.2.1]: https://github.com/Skycoder42/dart_pre_commit/compare/v5.2.0+1...v5.2.1
 [5.2.0+1]: https://github.com/Skycoder42/dart_pre_commit/compare/v5.2.0...v5.2.0+1
 [5.2.0]: https://github.com/Skycoder42/dart_pre_commit/compare/v5.1.0+1...v5.2.0